invitation update5

+++ src/app/api/workspace/[workspaceId]/invitation/[invitationId]/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { getCurrentUser } from '@/app/actions/getCurrentUser';
-import prisma from '@/lib/prismadb'; // Prisma Client
+import prisma from '@/lib/prismadb';
 
 export async function DELETE(
   req: NextRequest,
@@ -25,31 +25,55 @@ export async function DELETE(
   }
 
   try {
-    // Cek apakah undangan ada
-    const invitation = await prisma.invitation.findUnique({
-      where: { id: invitationId, workspaceId: workspaceId },
+    // Cek peran pengguna di workspace
+    const workspaceUser = await prisma.workspaceMember.findUnique({
+      where: {
+        userId_workspaceId: { userId: currentUser.id, workspaceId },
+      },
     });
 
-    if (!invitation) {
+    if (!workspaceUser) {
       return NextResponse.json(
-        { error: 'Invitation not found' },
+        { message: 'User not found in workspace' },
         { status: 404 }
       );
     }
 
-    // Hapus undangan dari database
-    await prisma.invitation.delete({ where: { id: invitationId } });
+    const isSuperAdmin = workspaceUser.role === 'SUPER_ADMIN';
+    const isAdmin = workspaceUser.role === 'ADMIN';
 
-    return (
-      NextResponse.json({
-        status: 'success',
-        message: 'Invitation revoked successfully',
-      }),
-      { status: 200 }
+    // Cek apakah undangan ada
+    const invitation = await prisma.invitation.findUnique({
+      where: { id: invitationId, workspaceId },
+    });
+
+    if (!invitation) {
+      return NextResponse.json(
+        { message: 'Invitation not found' },
+        { status: 404 }
+      );
+    }
+
+    // Validasi hak akses untuk mencabut undangan
+    if (
+      isSuperAdmin ||
+      (isAdmin && invitation.invitedById === currentUser.id)
+    ) {
+      await prisma.invitation.delete({ where: { id: invitationId } });
+      return NextResponse.json(
+        { message: 'Invitation revoked successfully' },
+        { status: 200 }
+      );
+    }
+
+    return NextResponse.json(
+      { message: 'You do not have permission to revoke this invitation' },
+      { status: 403 }
     );
   } catch (error) {
+    console.error(error);
     return NextResponse.json(
-      { error: 'Internal Server Error' },
+      { message: 'Internal Server Error' },
       { status: 500 }
     );
   }
